Requirement 3 - Upload of supporting documents
Lecturers are now able to upload their supporting documents when submitting a claim - limited to 5MB and only .pdf, .docx and .xlms file types.

All users are able to see this uploaded document - will change to only admins can view.

Documents will be uploaded to a 'uploads' file - this is currently not working.