Upload without notes works!
I have rewritten a large section of the home controller code, as well as added additional code to see if the achieved goal would work.

My program is currently bloated with unnecessary code - especially in the HomeController, but thats an issue to be fixed at another stage.

A lecturer is now able to upload a claim without adding any notes and the claim will go through with the notes section stating "No additional notes"!