Additional CSS applied - and upload works!
I have added slighly more css to the application.

I dont know what I did, but the application restarted differently than it has previously, and this finally fixed my uploads section, so a lecturers documents actually uploads to file! All be it under a random file name (Which to an extend helps because if 2 files have the same name the older one will get overwritten) - might still try to change it but Ill see.