Attempting to add login system
The login system I originally had I am replacing, Ive tried to enable hashing on passwords, but I have yet to come right and am going to try a more simple login solution. Currently this solution uses the database to store users, it seems to be working but does not redirect me to the correct pages.